docker/manager/requirements.txt
djangorestframework
psycopg2-binary
Pillow
